+++ quill_rag/reranker.py
@@ -3,9 +3,8 @@
 
 from __future__ import annotations
 
-import logging
 
-logger = logging.getLogger(__name__)
+from astrbot.api import logger
 
 
 class QuillReranker: